+++ src/mocks/UsageStatuses.json
@@ -9,7 +9,7 @@
   {
     "id": "1",
     "value": "In Use",
-    "code": "in_use",
+    "code": "in-use",
     "created_time": "2024-01-01T12:00:00.000+00:00",
     "modified_time": "2024-01-02T13:10:10.000+00:00"
   },
